Remove more deprecated things
