+++ sdk/typescript/src/types/index.ts
@@ -3,14 +3,6 @@
 
 export {
 	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	ObjectOwner,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	ProtocolConfig,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiJsonValue,
-} from './common.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
 	CheckpointedObjectId,
 	/** @deprecated Import type from `@mysten/sui.js/client` instead */
 	DisplayFieldsBackwardCompatibleResponse,
@@ -103,142 +95,6 @@ export {
 	/** @deprecated This method will be removed in a future version of the SDK */
 	isSuiObjectResponse,
 } from './objects.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	EventId,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	type MoveEventField,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	PaginatedEvents,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiEvent,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	type SuiEventFilter,
-	getEventPackage,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getEventSender,
-	/** @deprecated This method will be removed in a future version of the SDK */
-} from './events.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	BalanceChange,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	DevInspectResults,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	DryRunTransactionBlockResponse,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	type ExecuteTransactionRequestType,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	ExecutionStatus,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	ExecutionStatusType,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	Genesis,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	MoveCallSuiTransaction,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	OwnedObjectRef,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	PaginatedTransactionResponse,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	ProgrammableTransaction,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiArgument,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiCallArg,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiChangeEpoch,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiConsensusCommitPrologue,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectChange,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectChangeCreated,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectChangeDeleted,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectChangeMutated,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectChangePublished,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectChangeTransferred,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectChangeWrapped,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiTransaction,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiTransactionBlock,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiTransactionBlockData,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiTransactionBlockResponse,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiTransactionBlockResponseOptions,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	type SuiTransactionBlockResponseQuery,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	TransactionEffects,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	TransactionEffectsModifiedAtVersions,
-	/** @deprecated Use SuiEvent[] from `@mysten/sui.js/client` instead */
-	TransactionEvents,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	type TransactionFilter,
-	/** @deprecated This type will be removed in a future version of the SDK */
-	type EmptySignInfo,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getChangeEpochTransaction,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getConsensusCommitPrologueTransaction,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getCreatedObjects,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getEvents,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getExecutionStatus,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getExecutionStatusError,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getExecutionStatusGasSummary,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getExecutionStatusType,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getGasData,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getNewlyCreatedCoinRefsAfterSplit,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectChanges,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getProgrammableTransaction,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getPublishedObjectChanges,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTimestampFromTransactionResponse,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTotalGasUsed,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTotalGasUsedUpperBound,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransaction,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransactionDigest,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransactionEffects,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransactionGasBudget,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransactionGasObject,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransactionGasPrice,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransactionKind,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransactionKindName,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransactionSender,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getTransactionSignature,
-} from './transactions.js';
 
 export {
 	/** @deprecated Import type from `@mysten/sui.js/client` instead */
@@ -282,98 +138,3 @@ export {
 	/** @deprecated This method will be removed in a future version of the SDK */
 	extractStructTag,
 } from './normalized.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	Apy,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	Balance,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	CommitteeInfo,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	DelegatedStake,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	StakeObject,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiSystemStateSummary,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiValidatorSummary,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	Validators,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	ValidatorsApy,
-} from './validator.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	CoinBalance,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	CoinStruct,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	CoinSupply,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	PaginatedCoins,
-} from './coin.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	EndOfEpochInfo,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	EpochInfo,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	EpochPage,
-} from './epochs.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	type Unsubscribe,
-} from './subscriptions.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	ResolvedNameServiceNames,
-} from './name-service.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	DynamicFieldInfo,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	DynamicFieldName,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	DynamicFieldPage,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	DynamicFieldType,
-} from './dynamic_fields.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	Checkpoint,
-	/** @deprecated Current type is an alias for `any`, use `unknown` instead */
-	CheckpointCommitment,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	CheckpointPage,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	EndOfEpochData,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	GasCostSummary,
-} from './checkpoints.js';
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	AddressMetrics,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	AllEpochsAddressMetrics,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	NetworkMetrics,
-} from './metrics.js';
-
-export {
-	/** @deprecated This method will be removed in a future version of the SDK */
-	Contents,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	ContentsFields,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	ContentsFieldsWithdraw,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	DelegationStakingPool,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	DelegationStakingPoolFields,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	StakeSubsidy,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	StakeSubsidyFields,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	SuiSupplyFields,
-} from './validator.js';
