+++ sdk/typescript/test/unit/cryptography/secp256r1-publickey.test.ts
@@ -47,7 +47,6 @@ describe('Secp256r1PublicKey', () => {
 		let pub_key_base64 = toB64(pub_key);
 		const key = new Secp256r1PublicKey(pub_key_base64);
 		expect(key.toBase64()).toEqual(pub_key_base64);
-		expect(key.toString()).toEqual(pub_key_base64);
 	});
 
 	it('toBuffer', () => {
